Move data folder into core
