+++ packages/data-mate/src/builder/types/AnyBuilder.ts
@@ -1,4 +1,4 @@
-import { WritableData } from '../../data';
+import { WritableData } from '../../core';
 import { VectorType } from '../../vector';
 import { Builder, BuilderOptions } from '../Builder';
