+++ packages/data-mate/src/builder/types/BigIntBuilder.ts
@@ -1,5 +1,5 @@
 import { toBigIntOrThrow } from '@terascope/utils';
-import { WritableData } from '../../data';
+import { WritableData } from '../../core';
 import { VectorType } from '../../vector';
 import { Builder, BuilderOptions } from '../Builder';
