+++ packages/data-mate/src/builder/types/FloatBuilder.ts
@@ -1,5 +1,5 @@
 import { toFloatOrThrow } from '@terascope/utils';
-import { WritableData } from '../../data';
+import { WritableData } from '../../core';
 import { VectorType } from '../../vector';
 import { Builder, BuilderOptions } from '../Builder';
