+++ packages/data-mate/src/builder/types/IntBuilder.ts
@@ -1,6 +1,6 @@
 import { FieldType } from '@terascope/types';
 import { toIntegerOrThrow } from '@terascope/utils';
-import { WritableData } from '../../data';
+import { WritableData } from '../../core';
 import { VectorType } from '../../vector';
 import { Builder, BuilderOptions } from '../Builder';
