+++ packages/data-mate/src/core/WritableData.ts
