+++ packages/data-mate/src/core/index.ts
@@ -1,3 +1,6 @@
 export * from './config';
-export * from './data-helpers';
 export * from './DateValue';
+export * from './interfaces';
+export * from './ReadableData';
+export * from './utils';
+export * from './WritableData';
