+++ packages/data-mate/src/core/interfaces.ts
@@ -1,3 +1,9 @@
+export const HASH_CODE_SYMBOL = Symbol('__hash__');
+
+export const MAX_8BIT_INT = (2 ** 8) - 1;
+export const MAX_16BIT_INT = (2 ** 16) - 1;
+export const MAX_32BIT_INT = (2 ** 32) - 1;
+
 export type TypedArray = Uint8Array
 |Uint16Array
 |Uint32Array
