+++ packages/data-mate/src/core/utils.ts
@@ -1,7 +1,33 @@
 import { createHash } from 'crypto';
 import { getTypeOf, isFunction, toString } from '@terascope/utils';
+import {
+    HASH_CODE_SYMBOL,
+    MAX_16BIT_INT, MAX_32BIT_INT, MAX_8BIT_INT,
+    TypedArray, TypedArrayConstructor
+} from './interfaces';
 
-export const HASH_CODE_SYMBOL = Symbol('__hash__');
+/**
+ * Gets the correctly sized TypeArray depending on the length of items
+ */
+export function getPointerArray(size: number): TypedArray {
+    const PointerArray = getTypedArrayClass(size);
+    return new PointerArray(size);
+}
+
+/**
+ * Gets the correctly sized TypeArray constructor depending on the size of values being stored
+ */
+export function getTypedArrayClass(size: number): TypedArrayConstructor {
+    const maxIndex = size - 1;
+
+    if (maxIndex <= MAX_8BIT_INT) return Uint8Array;
+
+    if (maxIndex <= MAX_16BIT_INT) return Uint16Array;
+
+    if (maxIndex <= MAX_32BIT_INT) return Uint32Array;
+
+    return Float64Array;
+}
 
 export function md5(value: string|Buffer): string {
     return createHash('md5').update(value).digest('hex');
