--- packages/data-mate/src/data/index.ts
@@ -1,4 +1,0 @@
-export * from './interfaces';
-export * from './ReadableData';
-export * from './utils';
-export * from './WritableData';
