+++ packages/data-mate/src/vector/types/DateVector.ts
@@ -1,7 +1,6 @@
 import { Vector, VectorOptions } from '../Vector';
 import { VectorType } from '../interfaces';
-import { DateValue } from '../../core';
-import { ReadableData } from '../../data';
+import { DateValue, ReadableData } from '../../core';
 
 export class DateVector extends Vector<DateValue> {
     static valueToJSON({ value, formatted }: DateValue): string|number {
