+++ packages/data-mate/src/vector/types/FloatVector.ts
@@ -1,6 +1,6 @@
 import { Vector, VectorOptions } from '../Vector';
 import { VectorType } from '../interfaces';
-import { ReadableData } from '../../data';
+import { ReadableData } from '../../core';
 
 export class FloatVector extends Vector<number> {
     constructor(options: VectorOptions<number>) {
