+++ packages/data-mate/src/vector/types/IPRangeVector.ts
@@ -1,6 +1,6 @@
 import { Vector, VectorOptions } from '../Vector';
 import { VectorType } from '../interfaces';
-import { ReadableData } from '../../data';
+import { ReadableData } from '../../core';
 
 export class IPRangeVector extends Vector<string> {
     constructor(options: VectorOptions<string>) {
